Rename Terms.tsx TermsContainer.tsx
