+++ src/components/TermsContainer.tsx
